refactor: fold limrun preinstall into allocation
